requirements.txt
channels==4.2.2
Django==5.1.7
daphne==4.1.2
pymongo==4.11.3
channels_redis==4.2.1
redis==5.2.1
PyJWT==2.10.1